[agent][process-agent][dca] Split common config command to own package
- Follow up to https://github.com/DataDog/datadog-agent/pull/8912.
- Split config command to its own package to avoid dependency on `pkg/collector/python`.
- Addresses the issue with broken Windows build (`windows_zip_agent_binaries_x64` job).
