+++ cmd/agent/app/config.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/cmd/agent/app/settings"
 	"github.com/DataDog/datadog-agent/cmd/agent/common"
-	"github.com/DataDog/datadog-agent/cmd/agent/common/commands"
+	cmdconfig "github.com/DataDog/datadog-agent/cmd/agent/common/commands/config"
 	"github.com/DataDog/datadog-agent/pkg/api/util"
 	"github.com/DataDog/datadog-agent/pkg/config"
 	commonsettings "github.com/DataDog/datadog-agent/pkg/config/settings"
@@ -19,7 +19,7 @@ import (
 )
 
 func init() {
-	AgentCmd.AddCommand(commands.Config(getSettingsClient))
+	AgentCmd.AddCommand(cmdconfig.Config(getSettingsClient))
 }
 
 func setupConfig() error {
