+++ cmd/cluster-agent/app/config.go
@@ -11,7 +11,7 @@ import (
 	"fmt"
 
 	"github.com/DataDog/datadog-agent/cmd/agent/common"
-	"github.com/DataDog/datadog-agent/cmd/agent/common/commands"
+	cmdconfig "github.com/DataDog/datadog-agent/cmd/agent/common/commands/config"
 	"github.com/DataDog/datadog-agent/pkg/api/util"
 	"github.com/DataDog/datadog-agent/pkg/config"
 	commonsettings "github.com/DataDog/datadog-agent/pkg/config/settings"
@@ -21,7 +21,7 @@ import (
 )
 
 func init() {
-	ClusterAgentCmd.AddCommand(commands.Config(getSettingsClient))
+	ClusterAgentCmd.AddCommand(cmdconfig.Config(getSettingsClient))
 }
 
 func setupConfig() error {
